fix compilation warnings

+++ belle-sip/src/belle_sip_headers_impl.c
@@ -449,7 +449,7 @@ void belle_sip_header_user_agent_add_product(belle_sip_header_user_agent_t* user
 	user_agent->products = belle_sip_list_append(user_agent->products ,belle_sip_strdup(product));
 }
 int belle_sip_header_user_agent_get_products_as_string(const belle_sip_header_user_agent_t* user_agent,char* value,unsigned int value_size) {
-	unsigned int result = 0;
+	size_t result = 0;
 	belle_sip_list_t* list = user_agent->products;
 	for(;list!=NULL;list=list->next){
 		belle_sip_snprintf(value,value_size,&result,"%s ",(const char *)list->data);
